estruturas e size_t

+++ src-adaptative/adaptative-par.h
@@ -2,12 +2,12 @@
 #define ADAPTATIVEPAR_H_
 
 double adpt_parallel_for_and_reduce(
-    double (*kernel)(void*, int, int), void* data, int first, int last, 
+    double (*kernel)(void*, size_t, size_t), void* data, size_t first, size_t last, 
     double start_value, double (*rdct_fun)(double, double)
 );
 
 void adapt_parallel_for(
-    void (*kernel)(void*, int, int), void* data, int first, int last
+    void (*kernel)(void*, size_t, size_t), void* data, size_t first, size_t last
 );
 
 #endif //ADAPTATIVEPAR_H_
